src/requirements.txt
albumentations==2.0.8
ensemble_boxes==1.0.9
loguru==0.5.3
matplotlib==3.10.6
numba==0.61.2
numpy==1.21.2
opencv_python==4.6.0.66
PyYAML==6.0.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.60.0
